requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
llamaapi==0.1.36
multidict==6.0.5
nest-asyncio==1.6.0
psycopg2-binary==2.9.9
pymongo==4.8.0
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
yarl==1.9.4
